+++ build.mjs
@@ -86,7 +86,6 @@ await build({
   format: "cjs",
   external: ["obsidian"],
   alias: sourceAliases,
-  loader: { ".md": "text" },
   outfile: "packages/obsidian/dist/main.js",
   logLevel: "info",
 });